smartmail_offline/requirements.txt
streamlit>=1.32.0
spacy>=3.7.0
transformers>=4.38.0
torch>=2.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.5.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
openai>=1.14.0
